+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 ### [Unreleased]
 
+##### Added
+- Choose type of client on creation.
+- Implement Proof Key for Code Exchange by OAuth Public Clients.
+- Support for prompt parameter.
+
+##### Fixed
+- Not auto-approve requests for non-confidential clients (publics).
+
 ### [0.3.1] - 2016-03-09
 
 ##### Fixed